# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"User Service","description":"HTTP service for managing user info","version":""},"host":"localhost:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/user":{"get":{"tags":["userMethod"],"summary":"show userMethod","description":"Show user info","operationId":"userMethod#show","parameters":[{"name":"view","in":"query","description":"View to render","required":false,"type":"string","enum":["default","tiny"]},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodShowResponseBody"}},"404":{"description":"Not Found response.","schema":{"$ref":"#/definitions/UserMethodShowNotFoundResponseBody","required":["message","id"]}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/user/activate/{code}":{"get":{"tags":["userMethod"],"summary":"activate userMethod","description":"activate","operationId":"userMethod#activate","parameters":[{"name":"code","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodActivateResponseBody","required":["code"]}}},"schemes":["http"]}},"/user/change/email":{"post":{"tags":["userMethod"],"summary":"changeEmail userMethod","description":"Add n number of bottles and return their IDs. This is a multipart request and each part has field name 'bottle' and contains the encoded bottle info to be added.","operationId":"userMethod#changeEmail","parameters":[{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"ChangeEmailRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodChangeEmailRequestBody","required":["email"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodChangeEmailResponseBody","required":["code"]}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/user/change/info":{"post":{"tags":["userMethod"],"summary":"changeInfo userMethod","description":"Add new bottle and return its ID.","operationId":"userMethod#changeInfo","parameters":[{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"ChangeInfoRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodChangeInfoRequestBody","required":["name","icon"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodChangeInfoResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/user/change/password":{"post":{"tags":["userMethod"],"summary":"changePassword userMethod","description":"Remove bottle from storage","operationId":"userMethod#changePassword","parameters":[{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"ChangePasswordRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodChangePasswordRequestBody","required":["oldPassword","newPassword"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodChangePasswordResponseBody","required":["code"]}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/user/forgot/password":{"post":{"tags":["userMethod"],"summary":"forgotPassword userMethod","description":"Rate bottles by IDs","operationId":"userMethod#forgotPassword","parameters":[{"name":"ForgotPasswordRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodForgotPasswordRequestBody","required":["code","newPassword","email"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodForgotPasswordResponseBody","required":["code"]}}},"schemes":["http"]}},"/user/login":{"post":{"tags":["userMethod"],"summary":"login userMethod","description":"Creates a valid JWT","operationId":"userMethod#login","parameters":[{"name":"LoginRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodLoginRequestBody","required":["email","password"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodLoginResponseBody","required":["jwt"]}}},"schemes":["http"]}},"/user/register":{"post":{"tags":["userMethod"],"summary":"register userMethod","description":"register user","operationId":"userMethod#register","parameters":[{"name":"RegisterRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodRegisterRequestBody","required":["email","password"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodRegisterResponseBody","required":["code"]}}},"schemes":["http"]}},"/user/send/code":{"post":{"tags":["userMethod"],"summary":"sendVerifyCode userMethod","description":"send verify code to email","operationId":"userMethod#sendVerifyCode","parameters":[{"name":"SendVerifyCodeRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/UserMethodSendVerifyCodeRequestBody","required":["email"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/UserMethodSendVerifyCodeResponseBody","required":["code"]}}},"schemes":["http"]}}},"definitions":{"UserMethodActivateResponseBody":{"title":"UserMethodActivateResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Commodi impedit et dolorem velit explicabo.":"Sequi quo facere et sit expedita.","Maiores rem voluptatem temporibus.":"Est minima deserunt earum quam ex.","Sapiente quas unde quia vitae.":"Est quibusdam."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"Omnis veritatis fuga eum.":"Et et.","Sequi harum maxime inventore harum et esse.":"Est ea."},"message":"bad request"},"required":["code"]},"UserMethodChangeEmailRequestBody":{"title":"UserMethodChangeEmailRequestBody","type":"object","properties":{"email":{"type":"string","example":"Dolor et sit expedita repudiandae veniam."}},"example":{"email":"Nisi consequuntur sint."},"required":["email"]},"UserMethodChangeEmailResponseBody":{"title":"UserMethodChangeEmailResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Architecto est reiciendis consequatur ut eos qui.":"Voluptate nesciunt et.","Asperiores quis est modi.":"Quae magni.","Magnam consequatur corporis alias ex.":"Tempora labore dolores ad."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"Doloremque omnis corrupti aliquid.":"Autem sed eum.","Neque sint voluptas error incidunt aut repellendus.":"Quisquam ratione sed."},"message":"bad request"},"required":["code"]},"UserMethodChangeInfoRequestBody":{"title":"UserMethodChangeInfoRequestBody","type":"object","properties":{"icon":{"type":"string","description":"icon of user","example":"Alias repellat et ipsam."},"name":{"type":"string","description":"name of user","example":"Reiciendis quia velit quia est autem."}},"example":{"icon":"Dolorem error qui.","name":"Eius ad officia corrupti soluta qui voluptas."},"required":["name","icon"]},"UserMethodChangeInfoResponseBody":{"title":"Mediatype identifier: application/vnd.cellar.stored-bottle; view=default","type":"object","properties":{"activate":{"type":"boolean","example":false},"email":{"type":"string","description":"Vintage of bottle","example":"zcj","maxLength":100},"icon":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"id":{"type":"integer","description":"ID is the unique id of the bottle.","example":1,"format":"int64"},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100}},"description":"ChangeInfoResponseBody result type (default view)","example":{"activate":true,"email":"6al","icon":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":1,"name":"Blue's Cuvee"},"required":["id","name","icon","email"]},"UserMethodChangePasswordRequestBody":{"title":"UserMethodChangePasswordRequestBody","type":"object","properties":{"newPassword":{"type":"string","description":"new password","example":"new password"},"oldPassword":{"type":"string","description":"old password","example":"old password"}},"example":{"newPassword":"new password","oldPassword":"old password"},"required":["oldPassword","newPassword"]},"UserMethodChangePasswordResponseBody":{"title":"UserMethodChangePasswordResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Delectus vitae qui optio aut.":"Accusantium consequuntur qui enim omnis ducimus sit."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"A dolorum saepe eum hic.":"Iste earum.","Laudantium dolorem ipsum ea assumenda repellat quasi.":"Accusamus quisquam qui dolore aut consequatur."},"message":"bad request"},"required":["code"]},"UserMethodForgotPasswordRequestBody":{"title":"UserMethodForgotPasswordRequestBody","type":"object","properties":{"code":{"type":"string","example":"1234"},"email":{"type":"string","example":"Eos debitis itaque eveniet qui."},"newPassword":{"type":"string","example":"Pariatur aspernatur quo beatae omnis et."}},"example":{"code":"1234","email":"Eos dignissimos dolorem nihil aut velit.","newPassword":"Sunt est dolorem nostrum quisquam labore."},"required":["code","newPassword","email"]},"UserMethodForgotPasswordResponseBody":{"title":"UserMethodForgotPasswordResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Et eligendi.":"Et minus odio voluptas eaque sed.","Sapiente unde aspernatur voluptas laboriosam alias.":"Et aut doloribus dolorum repellendus sit earum."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"Et odit id quae ipsam.":"Non aut perspiciatis.","Quam assumenda ex vitae ut.":"Veniam mollitia in."},"message":"bad request"},"required":["code"]},"UserMethodLoginRequestBody":{"title":"UserMethodLoginRequestBody","type":"object","properties":{"email":{"type":"string","description":"Username used to perform signin","example":"123@email.com"},"password":{"type":"string","description":"Password used to perform signin","example":"password"}},"description":"Credentials used to authenticate to retrieve JWT token","example":{"email":"123@email.com","password":"password"},"required":["email","password"]},"UserMethodLoginResponseBody":{"title":"UserMethodLoginResponseBody","type":"object","properties":{"jwt":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"example":{"jwt":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"required":["jwt"]},"UserMethodRegisterRequestBody":{"title":"UserMethodRegisterRequestBody","type":"object","properties":{"email":{"type":"string","description":"email used to perform login","example":"123@email.com"},"password":{"type":"string","description":"Password used to perform signin","example":"password"}},"example":{"email":"123@email.com","password":"password"},"required":["email","password"]},"UserMethodRegisterResponseBody":{"title":"UserMethodRegisterResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Beatae dignissimos libero vel.":"Rerum molestiae quo."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"Distinctio numquam.":"Qui enim in temporibus consequatur.","Qui repudiandae ullam voluptatibus aspernatur cupiditate.":"Sed ut magni dolores ipsam recusandae.","Sit quas debitis.":"Ullam qui ducimus sint aut vel."},"message":"bad request"},"required":["code"]},"UserMethodSendVerifyCodeRequestBody":{"title":"UserMethodSendVerifyCodeRequestBody","type":"object","properties":{"email":{"type":"string","example":"Consequuntur et est quibusdam sed."}},"example":{"email":"Molestiae tempora quibusdam."},"required":["email"]},"UserMethodSendVerifyCodeResponseBody":{"title":"UserMethodSendVerifyCodeResponseBody","type":"object","properties":{"code":{"type":"integer","description":"code","example":200,"format":"int64"},"data":{"type":"object","example":{"Quia veniam.":"Qui tempore dolor modi veritatis magnam iste."},"additionalProperties":true},"message":{"type":"string","description":"message","example":"bad request"}},"example":{"code":200,"data":{"Error soluta voluptatem eveniet autem.":"Non ex temporibus.","Nam quaerat tempora debitis sed.":"Voluptatum recusandae quae dignissimos esse nam.","Quidem aut dignissimos quia cupiditate dolorum incidunt.":"Eligendi ea nisi dolorem."},"message":"bad request"},"required":["code"]},"UserMethodShowNotFoundResponseBody":{"title":"UserMethodShowNotFoundResponseBody","type":"object","properties":{"id":{"type":"string","description":"ID of missing user","example":"Nulla et excepturi molestiae."},"message":{"type":"string","description":"Message of error","example":"user 1 not found"}},"description":"Bottle not found","example":{"id":"Optio impedit nobis.","message":"user 1 not found"},"required":["message","id"]},"UserMethodShowResponseBody":{"title":"Mediatype identifier: application/vnd.cellar.stored-bottle; view=default","type":"object","properties":{"activate":{"type":"boolean","example":true},"email":{"type":"string","description":"Vintage of bottle","example":"alw","maxLength":100},"icon":{"type":"string","description":"Description of bottle","example":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","maxLength":2000},"id":{"type":"integer","description":"ID is the unique id of the bottle.","example":1,"format":"int64"},"name":{"type":"string","description":"Name of bottle","example":"Blue's Cuvee","maxLength":100}},"description":"ShowResponseBody result type (default view)","example":{"activate":true,"email":"mzi","icon":"Red wine blend with an emphasis on the Cabernet Franc grape and including other Bordeaux grape varietals and some Syrah","id":1,"name":"Blue's Cuvee"},"required":["id","name","icon","email"]}},"securityDefinitions":{"jwt_header_Authorization":{"type":"apiKey","description":"Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes \"api:read\" and \"api:write\".","name":"Authorization","in":"header"}}}